extract and test debounce function
